refactor(base): Deleted withStyles HOC
BREAKING CHANGE: Delete `withStyles` export from `base` package
