--- packages/base/src/lib/withStyles.ts
@@ -1,3 +1,0 @@
-import { withStyles } from '../withStyles';
-
-export { withStyles };
